implement models services and controllers for entities

+++ src/main/java/com/online_shop/MAP_Labor_2_Spring/models/ProductSpec.java
@@ -1,7 +1,5 @@
 package com.online_shop.MAP_Labor_2_Spring.models;
 
-
-import com.fasterxml.jackson.annotation.JsonIgnore;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
@@ -17,7 +15,7 @@ public class ProductSpec {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     @ManyToOne
     @JoinColumn(name = "product_id", referencedColumnName = "id")
     private Product product;
